+++ src/compiler/slipshow.ml
@@ -289,27 +289,9 @@ let to_grace htbl_include er =
     er
 
 let to_grace units errors =
-  let primary_loc = function
-    | Diagnosis.DuplicateID { occurrences; _ } -> List.nth_opt occurrences 0
-    | MissingFile { locs; _ } -> List.nth_opt locs 0
-    | WrongType { loc_reason; _ } -> Some loc_reason
-    | InconsistentOption { loc1; _ } -> Some loc1
-    | ParsingWarnor { warnor = UnusedArgument { loc = ploc; _ }; loc }
-    | ParsingWarnor { warnor = Parsing_failure { loc = ploc; _ }; loc } ->
-        Some (Diagnosis.loc_of_ploc loc ploc)
-    | ParsingError { loc; _ }
-    | MissingID { loc; _ }
-    | UnknownAttribute { loc; _ }
-    | UnknownFrontmatterField { loc; _ }
-    | FrontmatterParsing { loc; _ }
-    | InvalidFrontmatterLine { loc }
-    | ChildrenClassWithValue { loc }
-    | Simple { loc; _ } ->
-        Some loc
-  in
   let compare_position a b =
     let open Cmarkit in
-    match (primary_loc a, primary_loc b) with
+    match (Diagnosis.primary_loc a, Diagnosis.primary_loc b) with
     | None, None -> 0
     | None, _ -> 1
     | _, None -> -1